// src/app/models/user.model.ts
export interface User {
  balance_amount: string;
  cc: string;
  currency: string;
  email: string;
  fname: string;
  language: string;
  lname: string;
  mobile: string;
  name: string;
  pro_pic: string;
  rating: string;
  status: string;
  trips: string;
  accept_status: string;
  wallet: string;
}